test working cereal for storing double array into portable bianry
The header files of cereal were not capable of finding each other so I
have edited the header files so that the header files would be searching
other header files in relative positions. I have also created
testcereal.cpp to test a minimal example for my usecase of cereal
storing an array of doubles into binary and then loading the values
back. The test is successful.